fix: Change verify form, register form, fix RegEx, fix SCSS
