+++ src/layouts/Account/AccountInput.tsx
@@ -2,15 +2,22 @@ import { Form, Input } from 'antd';
 import { FC, memo } from 'react';
 import './index.scss';
 interface AccountInputProps {
-    type?: string | 'text';
+    type?: string;
     name: string;
     label: string;
-    placeholder?: string | 'Nhập vào trường này';
+    placeholder?: string;
     rules?: any;
 }
 
 const AccountInput: FC<AccountInputProps> = memo(
-    ({ type, name, label, placeholder, rules, ...props }) => {
+    ({
+        type = 'text',
+        name,
+        label,
+        placeholder = 'Nhập vào trường này',
+        rules,
+        ...props
+    }) => {
         return (
             <Form.Item
                 name={name}
@@ -24,7 +31,7 @@ const AccountInput: FC<AccountInputProps> = memo(
                 ]}
                 {...props}
             >
-                <Input type={type} placeholder={placeholder} />
+                <Input type={type} placeholder={placeholder} size="large" />
             </Form.Item>
         );
     }
