+++ src/layouts/Account/OverrideAntd.scss
@@ -4,8 +4,5 @@
     }
     .ant-form-item {
         margin-bottom: 8px;
-        .ant-input {
-            height: 40px !important;
-        }
     }
 }
