+++ src/views/Register/Inner.jsx
@@ -11,7 +11,7 @@ const Inner = memo(({ handleRegister }) => {
     const nameRules = useMemo(
         () => [
             {
-                pattern: /^[^0-9]*$/,
+                pattern: /^\D*$/,
                 message: 'Họ và tên không được chứa số.',
             },
         ],
@@ -28,37 +28,34 @@ const Inner = memo(({ handleRegister }) => {
         []
     );
 
-    const passwordRules = useMemo(
-        () => [
-            {
-                pattern: /^(?=.*\d)(?=.*\D)[^\s]{8,}$/,
-                message: 'Mật khẩu phải có ít nhất 8 ký tự gồm chữ và số.',
-            },
-        ],
-        []
-    );
+    // const passwordRules = useMemo(
+    //     () => [
+    //         {
+    //             pattern: /^(?=.*\d)(?=.*\D)[^\s]{8,}$/,
+    //             message: 'Mật khẩu phải có ít nhất 8 ký tự gồm chữ và số.',
+    //         },
+    //     ],
+    //     []
+    // );
 
-    const confirmPasswordRules = useMemo(
-        () => [
-            ...passwordRules,
-            ({ getFieldValue }) => ({
-                validator(_, value) {
-                    if (!value || getFieldValue('password') === value) {
-                        return Promise.resolve();
-                    }
-                    return Promise.reject(
-                        new Error('Mật khẩu nhập lại không trùng khớp.')
-                    );
-                },
-            }),
-        ],
-        [passwordRules]
-    );
+    // const confirmPasswordRules = useMemo(
+    //     () => [
+    //         ...passwordRules,
+    //         ({ getFieldValue }) => ({
+    //             validator(_, value) {
+    //                 if (!value || getFieldValue('password') === value) {
+    //                     return Promise.resolve();
+    //                 }
+    //                 return Promise.reject(
+    //                     new Error('Mật khẩu nhập lại không trùng khớp.')
+    //                 );
+    //             },
+    //         }),
+    //     ],
+    //     [passwordRules]
+    // );
     const handleFinish = useCallback(
         values => {
-            //Delete confirm password field before calling API
-            delete values.confirmPassword;
-            console.log(values);
             handleRegister(values);
         },
         [handleRegister]
@@ -92,21 +89,6 @@ const Inner = memo(({ handleRegister }) => {
                     type="text"
                     placeholder="Email"
                     rules={emailRules}
-                />
-                <AccountInput
-                    name="password"
-                    label="Mật khẩu"
-                    type="password"
-                    placeholder="Mật khẩu"
-                    tooltip="Mật khẩu phải có ít nhất 8 kí tự gồm chữ và số."
-                    rules={passwordRules}
-                />
-                <AccountInput
-                    name="confirmPassword"
-                    label="Nhập lại mật khẩu"
-                    type="password"
-                    placeholder="Nhập lại mật khẩu"
-                    rules={confirmPasswordRules}
                 />
                 <Button
                     className="account-layout-button"
