+++ src/views/Register/index.jsx
@@ -11,10 +11,9 @@ const Wrapper = memo(() => {
             const response = await userService.register(data);
             if (response.isSuccess) {
                 Message.sendSuccess(
-                    'Đăng ký thành công. Vui lòng xác thực email.'
+                    'Đăng ký thành công. Vui lòng kiểm tra email để xác thực tài khoản.'
                 );
                 navigate(routeConstants.VERIFY_EMAIL);
-                localStorage.setItem('verifyEmail', data.email);
             }
             return response;
         },
