+++ src/views/VerifyAccount/index.jsx
@@ -1,0 +1,35 @@
+import Message from 'components/Message';
+import { memo, useCallback, useState, useEffect } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
+import routeConstants from 'route/routeConstant';
+import userService from 'services/userService';
+import Inner from 'views/VerifyAccount/Inner';
+
+const Wrapper = memo(() => {
+    const navigate = useNavigate();
+    const [loading, setLoading] = useState(false);
+    const { token } = useParams();
+    const handleVerifyAccount = useCallback(
+        async data => {
+            setLoading(true);
+            const response = await userService.verifyAccount(data, token);
+            setLoading(false);
+            if (response.isSuccess) {
+                Message.sendSuccess(
+                    'Xác thực tài khoản thành công. Vui lòng đăng nhập để tiếp tục.'
+                );
+                navigate(routeConstants.LOGIN);
+            }
+        },
+        [navigate, token]
+    );
+
+    return (
+        <Inner handleVerifyAccount={handleVerifyAccount} loading={loading} />
+    );
+});
+Wrapper.displayName = 'Verify Account';
+
+const VerifyAccount = Wrapper;
+
+export default VerifyAccount;
